test: Simplify and re-organize lifecycle helpers in integration/core
The idea is that each test helper file only exposes `before`,
`beforeEach`, `after`, and `afterEach`, without using namespaces such as
`helpers.jellyfish`, `helpers.cache`, etc that make everything more
confusing than it should be.

Change-type: patch
