+++ Makefile
@@ -333,8 +333,7 @@ test-integration:
 	FILES="'./test/integration/**/*.spec.js'" make test
 
 test-e2e:
-	FILES="'./test/e2e/**/*.spec.{js,jsx}'" \
-		AVA_OPTS="--serial" make test
+	FILES="'./test/e2e/**/*.spec.{js,jsx}'" make test
 
 test-unit-%:
 	FILES="'./test/unit/$(subst test-unit-,,$@)/**/*.spec.{js,jsx}'" SCRUB=0 make test
@@ -357,8 +356,7 @@ test-integration-%:
 	FILES="'./test/integration/$(subst test-integration-,,$@)/**/*.spec.js'" make test
 
 test-e2e-%:
-	FILES="'./test/e2e/$(subst test-e2e-,,$@)/**/*.spec.{js,jsx}'" \
-		AVA_OPTS="--serial" make test
+	FILES="'./test/e2e/$(subst test-e2e-,,$@)/**/*.spec.{js,jsx}'" make test
 
 ngrok-%:
 	ngrok start -config ./ngrok.yml $(subst ngrok-,,$@)
